+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [x] - Next
+
+- Fix alignment issue
+- Fix positioning issues
+- Change width and height to be defined on animations paint time (breaking).
+
+
 ## [0.0.1] - Initial release
 
 - Initial implementation